schemastore: share create view query normalizer
